Remove Selenium dependency
The new code uses the (internal) BGG JSON API directly.
This removes many dependencies, speeds up the process, makes it more
reliable and makes it possible to handle several copies of a game at
the same time.

When "collid" is empty or --force-new is given, the importer creates
new copies of the games. Without --force-new and with non-empty
"collid" matching existing copies of the games will be updated.